+++ src/echotools/__init__.py
@@ -8,11 +8,8 @@ from __future__ import annotations
 import importlib
 from typing import Any
 
-from echotools.compat import install_compat_aliases
 from echotools.plat.sdk import EchoTools
 from echotools.version import get_version
-
-install_compat_aliases()
 
 __version__ = get_version()